@@ -1,5 +1,5 @@
 class AuthenticationController < ApplicationController
-  skip_before_filter :authenticate_user!
+  skip_before_action :authenticate_user!
 
   def login
     redirect_to root_path
