@@ -1,4 +1,4 @@
-class CreateScripts < ActiveRecord::Migration
+class CreateScripts < ActiveRecord::Migration[4.2]
   def change
     create_table :scripts do |t|
       t.string :name
