@@ -1,4 +1,4 @@
-class CreateAttachments < ActiveRecord::Migration
+class CreateAttachments < ActiveRecord::Migration[4.2]
   def change
     create_table :attachments do |t|
       t.integer :script_id
