@@ -1,4 +1,4 @@
-class DropTableHooks < ActiveRecord::Migration
+class DropTableHooks < ActiveRecord::Migration[4.2]
   def change
     drop_table :hooks
   end
