@@ -39,7 +39,7 @@ describe LogsController, type: :controller do
   describe "GET show" do
     it "assigns the requested log as @log" do
       log = Log.create! valid_attributes
-      get :show, {:id => log.to_param}, valid_session
+      get :show, params: { id: log }, session: valid_session
       expect(controller.instance_variable_get("@log")).to eq(log)
     end
   end
